+++ routes/japan.js
@@ -84,7 +84,10 @@ router.post('/img', isLoggedIn, upload.single('img'), (req, res) => {
 
 router.delete('/:id', isLoggedIn, async (req, res, next) => {
     try{
-        await Japan.destroy({ where: { id: req.params.id }});
+        const japan = await Japan.findOne({ where: { id: req.params.id } });
+        if (japan.UserId === req.user.id) {
+            await Japan.destroy({ where: { id: req.params.id }});
+        }
     } catch (error) {
         console.error(error);
         next(error);
@@ -93,23 +96,25 @@ router.delete('/:id', isLoggedIn, async (req, res, next) => {
 
 router.patch('/posting/:id', isLoggedIn, async (req, res, next) => {
     try {
-        if(req.body.url) {
-            await Japan.update({
-                title: req.body.title,
-                content: req.body.content,
-                img: req.body.url,
-            }, {
-                where: { id: req.params.id },
-            });
-        } else {
-            await Japan.update({
-                title: req.body.title,
-                content: req.body.content,
-            }, {
-                where: { id: req.params.id },
-            });
+        const japan = await Japan.findOne({ where: { id: req.params.id } });
+        if (japan.UserId === req.user.id) {
+            if(req.body.url) {
+                await Japan.update({
+                    title: req.body.title,
+                    content: req.body.content,
+                    img: req.body.url,
+                }, {
+                    where: { id: req.params.id },
+                });
+            } else {
+                await Japan.update({
+                    title: req.body.title,
+                    content: req.body.content,
+                }, {
+                    where: { id: req.params.id },
+                });
+            }
         }
-        
     } catch (err) {
         console.error(err);
         next(err);
